refactor solver_jobs.py to avoid complaints from codeclimate
