+++ services/api-server/src/simcore_service_api_server/api/root.py
@@ -1,4 +1,5 @@
 from fastapi import APIRouter
+from simcore_service_api_server.api.routes import solvers_jobs_getters
 
 from ..core.settings import ApplicationSettings
 from .routes import (
@@ -27,6 +28,9 @@ def create_router(settings: ApplicationSettings):
     router.include_router(files.router, tags=["files"], prefix="/files")
     router.include_router(solvers.router, tags=["solvers"], prefix="/solvers")
     router.include_router(solvers_jobs.router, tags=["solvers"], prefix="/solvers")
+    router.include_router(
+        solvers_jobs_getters.router, tags=["solvers"], prefix="/solvers"
+    )
     router.include_router(studies.router, tags=["studies"], prefix="/studies")
     router.include_router(studies_jobs.router, tags=["studies"], prefix="/studies")
     router.include_router(wallets.router, tags=["wallets"], prefix="/wallets")
